Use getAvailableClients everywhere
- move getAvailableClients to its own file
- use getAvailableClients in configManager, listConnections and connectionCommand
