+++ src/roomCommand.ts
@@ -1,35 +1,9 @@
 import { program } from 'commander'
 import { setRoomInFocus } from './appCommands/setInFocus.js'
-import { addRoom, constructAndLoadRoom, getS3Credentials, listRooms } from './configManager.js'
+import { addRoom, constructAndLoadRoom, listRooms } from './configManager.js'
 import { createRoom } from './createRoom.js'
 import chalk from 'chalk'
-import { LocalClient } from '@diograph/local-client'
-import { S3Client } from '@diograph/s3-client'
-// TODO: Fix to '@diograph/diograph/types' (works in S3Client, at least ConnectionClient)
-import { ConnectionClientList } from '@diograph/diograph/types.js'
-
-const getAvailableClients = async (): Promise<ConnectionClientList> => {
-  const availableClients: ConnectionClientList = {
-    LocalClient: {
-      clientConstructor: LocalClient,
-    },
-  }
-
-  // S3Client is not available if no credentials found from config file
-  try {
-    const credentials = await getS3Credentials()
-    availableClients['S3Client'] = {
-      clientConstructor: S3Client,
-      credentials: { region: 'eu-west-1', credentials },
-    }
-  } catch (err) {
-    if ((err as Error).message !== 'No s3Credentials defined in config file') {
-      throw err
-    }
-  }
-
-  return availableClients
-}
+import { getAvailableClients } from './getAvailableClients.js'
 
 const exitIfRoomAlreadyExists = async (roomAddress: string, method?: string) => {
   const roomList = await listRooms()
